backend/requirements.txt
# Django
django==3.1.3
djangorestframework==3.12.2
markdown==3.3.3
django-cors-headers==3.5.0
djangorestframework-jsonapi==4.0.0
requests==2.25.0